Enable rewrite-to-brgemm to preserve parallelism at tensor level.
At tensor level `linalg::GenerateLoopNest` emits only scf.for loop even
if the loops have parallel semantics. This changes enable to rewrite
parallel scf.for to scf.forall in the same way we did for the tile and
fuse pass. The PRs mainly move code around to make this possible
exposing `populateScfForToForAllRewritePattern` as utils in
`TransformUtils.h`.
