+++ include/TPP/TransformUtils.h
@@ -78,6 +78,12 @@ bool isMatmulOp(Operation *op,
 bool hasMulAddBody(linalg::LinalgOp linalgOp,
                    SmallVectorImpl<Value> *capturedOperands = nullptr);
 
+// Rewrite scf.for to scf.forall. Assumes the loop to be parallel and
+// marked with `kLoopId`.
+constexpr const static llvm::StringLiteral kLoopId = "parallel";
+constexpr const static llvm::StringLiteral kLoopRoot = "root";
+void populateScfForToForAllRewritePattern(RewritePatternSet &patterns);
+
 } // namespace utils
 } // namespace linalgx
 } // namespace mlir
