+++ lib/TPP/CMakeLists.txt
@@ -39,6 +39,7 @@ add_mlir_library(MLIRTPP
 
   # Test Passes
     TestMatchers.cpp
+    TestForToForAllRewrite.cpp
 
     ADDITIONAL_HEADER_DIRS
     ${PROJECT_SOURCE_DIR}/include/TPP
