+++ lib/TPP/TileConsumerAndFuseProducers.cpp
@@ -55,8 +55,8 @@ struct ReplaceIterArgs : public OpRewritePattern<scf::ForOp> {
 
   LogicalResult matchAndRewrite(scf::ForOp forOp,
                                 PatternRewriter &rewriter) const override {
-    auto metadata = forOp->getAttrOfType<StringAttr>("fusion");
-    if (!metadata || metadata.getValue() != "root")
+    auto metadata = forOp->getAttrOfType<StringAttr>(linalgx::utils::kLoopId);
+    if (!metadata || metadata.getValue() != linalgx::utils::kLoopRoot)
       return failure();
     if (forOp.getNumRegionIterArgs() != 1)
       return failure();
@@ -65,91 +65,6 @@ struct ReplaceIterArgs : public OpRewritePattern<scf::ForOp> {
     if (nestedLoops.size() == 0)
       return failure();
     replaceIterArgs(rewriter, forOp, nestedLoops[nestedLoops.size() - 1]);
-    return success();
-  }
-};
-
-// Convert scf.for to scf.forall after fusion.
-struct ConvertToForAll : public OpRewritePattern<scf::ForOp> {
-  using OpRewritePattern<scf::ForOp>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(scf::ForOp forOp,
-                                PatternRewriter &rewriter) const override {
-    auto metadata = forOp->getAttrOfType<StringAttr>("fusion");
-    if (!metadata || metadata.getValue() != "root")
-      return failure();
-    if (forOp.getNumRegionIterArgs() != 1)
-      return failure();
-    SmallVector<scf::ForOp> nestedLoops;
-    getPerfectlyNestedLoops(nestedLoops, forOp);
-    if (nestedLoops.size() == 0)
-      return failure();
-
-    SmallVector<Value> loopArgs;
-    SmallVector<OpFoldResult> lbs, ubs, steps;
-    for (scf::ForOp &currentLoop : nestedLoops) {
-      if (currentLoop.getNumRegionIterArgs() != 1)
-        return failure();
-      loopArgs.push_back(currentLoop.getInductionVar());
-      lbs.push_back(currentLoop.getLowerBound());
-      ubs.push_back(currentLoop.getUpperBound());
-      steps.push_back(currentLoop.getStep());
-      auto metadata = currentLoop->getAttrOfType<StringAttr>("fusion");
-      if (metadata && metadata.getValue() == "tail")
-        loopArgs.push_back(currentLoop.getRegionIterArg(0));
-    }
-
-    rewriter.replaceOpWithNewOp<scf::ForallOp>(
-        forOp, lbs, ubs, steps, ValueRange{forOp.getInitArgs()},
-        /*mapping=*/std::nullopt,
-        [&](OpBuilder &nestedBuilder, Location loc, ValueRange regionArgs) {
-          IRMapping mapping;
-          assert(loopArgs.size() == regionArgs.size() &&
-                 "expect same region args");
-          mapping.map(loopArgs, regionArgs);
-          Block *innerLoopBlock = nestedLoops[nestedLoops.size() - 1].getBody();
-          auto yieldOp = cast<scf::YieldOp>(innerLoopBlock->getTerminator());
-          auto insertSlice =
-              yieldOp.getOperands()[0].getDefiningOp<tensor::InsertSliceOp>();
-          assert(insertSlice && "must be an insert slice");
-          for (auto &nestedOp : innerLoopBlock->without_terminator()) {
-            if (&nestedOp == insertSlice.getOperation()) {
-              auto term = nestedBuilder.create<scf::InParallelOp>(loc);
-              nestedBuilder.setInsertionPointToStart(term.getBody());
-              Value sourceVal = mapping.lookup(insertSlice.getSource());
-              Value destVal = mapping.lookup(insertSlice.getDest());
-              SmallVector<OpFoldResult> offsets;
-              for (OpFoldResult offset : insertSlice.getMixedOffsets()) {
-                if (auto valueOffset = offset.dyn_cast<Value>())
-                  offsets.push_back(mapping.lookup(valueOffset));
-                else
-                  offsets.push_back(offset);
-              }
-              SmallVector<OpFoldResult> sizes;
-              for (OpFoldResult size : insertSlice.getMixedSizes()) {
-                if (auto valueSize = size.dyn_cast<Value>())
-                  sizes.push_back(mapping.lookupOrDefault(valueSize));
-                else
-                  sizes.push_back(size);
-              }
-              SmallVector<OpFoldResult> strides;
-              for (OpFoldResult stride : insertSlice.getMixedStrides()) {
-                if (auto valueStride = stride.dyn_cast<Value>())
-                  strides.push_back(mapping.lookupOrDefault(valueStride));
-                else
-                  strides.push_back(stride);
-              }
-              assert(offsets.size() == sizes.size());
-              assert(offsets.size() == strides.size());
-
-              nestedBuilder.create<tensor::ParallelInsertSliceOp>(
-                  loc, sourceVal, destVal, offsets, sizes, strides);
-              continue;
-            }
-            Operation *clone = nestedBuilder.clone(nestedOp, mapping);
-            mapping.map(nestedOp.getResults(), clone->getResults());
-          }
-        });
     return success();
   }
 };
@@ -500,10 +415,9 @@ fuseWithEltwise(RewriterBase &rewriter, TilingInterface consumer,
     LLVM_DEBUG(llvm::dbgs() << "NEW OP: " << tiledOp << "\n");
   }
   if (!tilingResult->loops.empty()) {
-    int64_t last = tilingResult->loops.size() - 1;
-    tilingResult->loops[0]->setAttr("fusion", rewriter.getStringAttr("root"));
-    tilingResult->loops[last]->setAttr("fusion",
-                                       rewriter.getStringAttr("tail"));
+    tilingResult->loops[0]->setAttr(
+        linalgx::utils::kLoopId,
+        rewriter.getStringAttr(linalgx::utils::kLoopRoot));
   }
   tileAndFuseResult.loops = std::move(tilingResult->loops);
   for (const auto &result : llvm::enumerate(llvm::zip_equal(
@@ -703,7 +617,7 @@ struct TileConsumerAndFuseProducers
       // Patterns for scf.forall.
       RewritePatternSet patterns(&ctx);
       if (this->useForAll)
-        patterns.add<ConvertToForAll>(&ctx);
+        linalgx::utils::populateScfForToForAllRewritePattern(patterns);
       // Fold unit-extent dims for linalg on tensors.
       linalg::populateFoldUnitExtentDimsViaSlicesPatterns(patterns);
       tensor::populateMergeConsecutiveInsertExtractSlicePatterns(patterns);
